day_95/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
openai>=1.55.0
speechrecognition==3.10.0
pyttsx3==2.90
numpy>=1.26.0
pandas>=2.1.0
python-dotenv==1.0.0
requests==2.31.0
textblob==0.17.1
scikit-learn>=1.3.0
python-multipart==0.0.6
Werkzeug==2.3.7
httpx>=0.25.0 